Conditional removal of nautilus plugin la file, bug #343643.
Package-Manager: portage-2.2.0_alpha3/cvs/Linux x86_64

+++ app-arch/file-roller/file-roller-2.32.0.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2010 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/app-arch/file-roller/file-roller-2.32.0.ebuild,v 1.2 2010/10/20 20:15:18 eva Exp $
+# $Header: /var/cvsroot/gentoo-x86/app-arch/file-roller/file-roller-2.32.0.ebuild,v 1.3 2010/11/01 11:25:55 eva Exp $
 
 EAPI="3"
 GCONF_DEBUG="no"
@@ -50,8 +50,10 @@ src_prepare() {
 
 src_install() {
 	gnome2_src_install
-	find "${ED}"usr/$(get_libdir)/nautilus -name "*.la" -delete \
-		|| die "la file removal failed"
+	if use nautilus; then
+		find "${ED}"usr/$(get_libdir)/nautilus -name "*.la" -delete \
+			|| die "la file removal failed"
+	fi
 }
 
 pkg_postinst() {
